+++ app/src/main/java/com/fongmi/android/tv/utils/WebViewUtil.java
@@ -57,7 +57,7 @@ public class WebViewUtil {
         settings.setDomStorageEnabled(true);
         settings.setDatabaseEnabled(true);
         settings.setMediaPlaybackRequiresUserGesture(false);
-        settings.setMixedContentMode(WebSettings.MIXED_CONTENT_ALWAYS_ALLOW);
+        settings.setMixedContentMode(WebSettings.MIXED_CONTENT_NEVER_ALLOW);
         settings.setRenderPriority(WebSettings.RenderPriority.HIGH);
         logSettings(webView, role);
     }